Fixing minor issues, remove comments, from PR
