+++ autoprof/models/planesky_model.py
@@ -55,7 +55,7 @@ class Plane_Sky(Sky_Model):
             self["delta"].set_uncertainty([0.1, 0.1], override_locked=True)
 
     def evaluate_model(self, image, X = None, Y = None, **kwargs):
-        if X is None:
+        if X is None or Y is None:
             X, Y = image.get_coordinate_meshgrid_torch(
                 self["center"].value[0], self["center"].value[1]
             )
